xen/arm: vpl011: Add PL011 emulation support
At the moment, vpl011 driver provides support for SBSA UART emulation.
This requires the guest to have a driver aware of differences between
PL011 and SBSA, since the latter is only a subset of the former. PL011 is
the de-facto standard UART in Arm based systems and it is easier to find
the driver for it rather than for SBSA UART.

Add support for PL011 emulation based on revision r1p5 documented in
ARM DDI0183G:
 -> populate missing registers in struct vpl011,
 -> emulate missing registers:
     - most registers do not need special handling apart from keeping
       track of the values read/written. The ones that require extra steps
       are: PERIPH_ID{0-3}, CELL_ID{0-3}, LCR_H, IFLS.
 -> support handling different FIFO levels and the corresponding IRQs
    assertion, de-assertion,
 -> support handling the state where FIFO is disabled.

Similar to SBSA UART emulation, it is not always possible to follow the
UART specification 1:1. For example, we use larger buffers than PL011's
mandated 32-byte FIFO size for better performance and to avoid possible
loss of characters.

There are features that cannot really be emulated such as DMA, IrDA SIR,
modem control, HW flow control and the ones for which it is not worth it
such as loopback operation.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Reviewed-by: Ayan Kumar Halder <ayan.kumar.halder@amd.com>
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>