xen/eemi: Make PM_FEATURE_CHECK handling common
At the moment, we do not handle PM_FEATURE_CHECK call for zynqMP unlike
Versal and Versal-Net. This call is safe and applies to all the
platforms, so make the handling common by moving it to xilinx_eemi() next
to other safe calls. Drop the forward_to_fw block in versal_eemi() and
versal_net_eemi() given no use.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>