xen/eemi: Add PM_SECURE_AES handling
PM_SECURE_AES is used to access AES hardware to encrypt/decrypt the data
using AES-GCM core. At the moment, we don't handle this call which
results in printing a warning about unhandled PM call (only for domUs,
dom0 requests are forwarded to firmware as usual). Add a definition for
this PM API call and a handler in xilinx_eemi() next to other calls that
are exclusive to the hardware domain. For domUs, it will result in
printing "... No access" which is more informative.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>