xen/arm: Use host layout for hwdom if LLC coloring enabled
At the moment, if LLC coloring is enabled, memory for hwdom is allocated
using allocate_memory() and thus represents the typical Xen guest memory
map. In case of hardware domain, this can create conflicts with MMIO
regions that are mapped to it at a later stage (the same issue applies
to reserved memory regions). Fix it by undergoing following steps:
 - make a call to find_unallocated_memory() to retrieve host memory
   layout (RAM regions excluding reserved memory regions),
 - allocate memory from previously retrieved host RAM regions,
 - use domain_use_host_layout() instead of is_domain_direct_mapped()
   for grant table and extended regions (colored hwdom is no longer
   direct mapped).

Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Signed-off-by: Stefano Stabellini <stefano.stabellini@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>