arm/pci: Add pci-scan boot argument
This patch adds a Xen boot arguments that, if enabled, causes a call to
existing code to scan pci devices enumerated by the firmware.

This patch also makes an existing debug function viewable outside its
translation unit, and uses this to dump the PCI devices found.
The debug message is controlled by an additional configurable added to
arch/arm/Kconfig.debug.

Additionally, this patch modifies segment loading to ensure that PCI
devices on other segments are properly found.

This will be needed ahead of dom0less support for pci passthrough on
arm.

Issue-Id: SCM-1368
Signed-off-by: Luca Fancellu <luca.fancellu@arm.com>
Signed-off-by: Edward Pickup <Edward.Pickup@arm.com>
Change-Id: I79eb1325019fb37fe0c7951070efe5745fed6905
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>