xen/pci: helper function for hwdom using vpci
When pci-scan is enabled and Xen supports vpci for guests, Xen will
scan the pci bus to find devices and emulate the pci bus, so the hw
domain must see the emulated bus instead of the real one.

A new helper function, hwdom_uses_vpci, is implemented and returns
true when pci-scan is enabled and Xen is built with
CONFIG_HAS_VPCI_GUEST_SUPPORT=y.

Issue-Id: SCM-1012
Signed-off-by: Luca Fancellu <luca.fancellu@arm.com>
Change-Id: Ifa6e0f91bde0cf524e91f811c1ac30afa3ed5213
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>