xen/arm: pci device passthrough for dom0less guests
Introduce the xen,pci-assigned property for the passthrough
node that is used to assign pci devices to a dom0less guest.

When the property is found, the addressed pci device is
assigned to the guest and the vpci DT node is added to the
guest DT.

Update the documentation about this property.

Issue-Id: SCM-1012
Signed-off-by: Luca Fancellu <luca.fancellu@arm.com>
Change-Id: I568aea739c1731f6d925a0bfbe581581187e6301
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>