arm/pci: select HAS_PCI_MSI
Currently, there is some pdev data that is not getting properly
initialized (in particular pdev->msi_maxvec). As a result, when passing
through a PCI device with 2 or more MSI vectors to a domU, only the
first MSI vector gets signaled in the domU. When building with
PCI_PASSTHROUGH enabled, also select HAS_PCI_MSI to enable the proper
pdev state initialization in Xen. This resolves the issue with passing
through devices with 2 or more MSI vectors.

Signed-off-by: Stewart Hildebrand <stewart.hildebrand@amd.com>
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>