x86/boot: refactor bzimage parser to be re-entrant
The bzimage logic uses the unit global orig_image_len to hold the
original module length for the kernel when the headroom is calculated.
It then uses orig_image_len to locate the start of the bzimage when
the expansion is done. This is an issue when more than one bzimage
is processed by the headroom calculation logic, as it will leave
orig_image_len set to the length of the last bzimage it processed.

Remove the global, now that the information can be deduced from
boot_module.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>
Signed-off-by: Alejandro Vallejo <alejandro.garciavallejo@amd.com>
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>
Reviewed-by: Jason Andryuk <jason.andryuk@amd.com>