dt-overlay: Fix removal of dt overlays
At the moment, the attempt to remove the previously added overlay fails.
This is because in remove_nodes(), we call remove_all_{irqs,iomems} for
every node, even though they should be executed just once as they operate
on all the ranges right away. Furthermore, we try to call at the end
{irq,iomem}_remove_cb to remove access rights despite the same being
already done as part of previous remove_all_{irqs,iomems}. This is
redundant as we only need to do this step once.

Take the opportunity to print I/O ranges in hex instead of decimal.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>