dt-overlay: Allow removing devices behind IOMMU from dom0
At the moment, the attempt to remove device, previously added with an
overlay, that is behind IOMMU would fail because the device is still
assigned (i.e. domain_list is not empty). In order to allow such devices
to be fully removed we need to call iommu_deassign_dt_device() before
executing iommu_remove_dt_device().

Take the opportunity to fix yet another NULL ptr dereference.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>