xen/arm: dom0less delay xenstore initialization
To allocate the xenstore event channel and initialize the grant table
entry, the xenstore domid is neeed.  A dom0 is craeted before the domUs,
so it is normally available through hardware_domain.  With capabilities
and dom0less, the xenstore domain may not be created first.

Keep the population of the page and HVM_PARAM_STORE_PFN in the normal
domain construction, but delay event channel and seeding to after all
domUs are created.  HVM_PARAM_STORE_PFN now serves as indication to
setup xenstore since the device tree is no longer immediately available.
0 means no xenstore.  ~0ULL means legacy so only the event channel needs
setup, and any other value means to seed the page.

dom0 needs to set xs_domid when it is serving as the xenstore domain.

Signed-off-by: Jason Andryuk <jason.andryuk@amd.com>