xen/arm: update vGIC #address cells to 2
When exposing a vITS node to a guest, the vGIC node address-cells is 2.
Change the vGIC node's #address-cells to 2 in all cases, since there is
no harm in doing so unconditionally, even when not exposing a vITS node.
Update the virtio-pci node's interrupt-map to reflect the new
address-cells value. This resolves a conflict when both vPCI and
virtio-pci are enabled simultaneously.

Signed-off-by: Stewart Hildebrand <stewart.hildebrand@amd.com>