x86/domain-builder: Support "reg" props in xen,domain nodes' blobs
In exotic configurations it's possible to pass modules in specific
areas of the initial system memory. Allow modules to be described
by "reg" properties, and add those modules once parsed to the list of
modules in boot_info for tracking purposes, with care not to add the
same module multiple times.

Signed-off-by: Alejandro Vallejo <alejandro.garciavallejo@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>
Reviewed-by: Jason Andryuk <jason.andryuk@amd.com>