x86: Make create_dom0 use boot_domains' flags
Move the existing defaults from create_dom0 to the domain builder's
fallback path. Now create_dom0 behaves the same with or without booting
off a DTB.

The dom0_max_vcpus() call is required so that dom0_nodes and dom0_cpus are
populated. Memory management and the scheduler rely on them. The next patch
sets them to default values in order to ignore the commandline.

Signed-off-by: Alejandro Vallejo <alejandro.garciavallejo@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>
Reviewed-by: Jason Andryuk <jason.andryuk@amd.com>