tools/init-dom0less: Use introduced to determine no-enhanced
An untargetable domain cannot have its HVM params read, and
init-dom0less reads HVM_PARAM_STORE_EVTCHN to determines whether or not
a domain has xenstore.

xenstored had a similar issue, and it just tries to map a domain's
reserved grant, GNTTAB_RESERVED_XENSTORE, to see if it is accessible.
If successful, xenstored introduces the domain.  Use the existing
introduced state to determine if init-dom0less should try and read the
param.  If already read, initialization is skipped.

This allows for running init-dom0less from a xenstored|hardware domain.
It relies on the local socket xenstore connections being considered
privileged.

oxenstored has not been updated, so the exist code remains for backwards
compatibility.

Signed-off-by: Jason Andryuk <jason.andryuk@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>