xsm/dummy: Update XSM_HW_PRIV platform ops for safety
Use finer granulatity for platform_ops with an eye toward a safety
configuration.  ops for proper hardware operation are allowed to
hardware_domain.  ops that are for querying data are allow.  ops that
affect platform state are disallowed.

Counter-intuitively, XENPF_enter_acpi_sleep is not allowed for the
hardware domain, because we don't want the hardware_domain affecting the
overall system state - control domain should be responsible for that.

XENPF_efi_runtime_call is XSM_PRIV.  It could be split to allow
hardware_domain "read" access while control domain can perform "writes".
That is not implemented.

physdev_ops all remain XSM_HW_PRIV as they are hardware-centric.

Signed-off-by: Jason Andryuk <jason.andryuk@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>