libxl: use QEMU machine type 'xenpvh' for PVH guests
When qemu-dm is required, instead of 'xenpv', use the
'xenpvh' QEMU instance that is tailored for PVH guests.

In contrast to the 'xenpv' instance, 'xenpvh' instance
has smp support. So, qemu-dm for PVH guests is launched
with the number of vcpus specified via -smp, similarly
to HVM guests.

Introduce new fields in the domain_build_info struct
of PVH guests to store the RAM layout. Those fields
are eventually used to populate the corresponding
properties of the 'xenpvh' instance.

Signed-off-by: Xenia Ragiadakou <xenia.ragiadakou@amd.com>
Reviewed-by: Jason Andryuk <jason.andryuk@amd.com>
Reviewed-by: Edgar E. Iglesias <edgar.iglesias@amd.com>