libacpi: add PCI Root bridge at PCI Segment 1 in PVH DSDT
Add an ECAM PCI Root bridge in the DSDT of PVH guests.
This PCI Root bridge is intended to be used for the discovery
of device model emulated PCI devices.

The PCI Segment is 1 instead of 0, because Segment 0 would be
more suitable to be reserved for passed-through PCI devices,
that can be also discovered via legacy CF8/CFC IO ports.

Add new fields in ACPI info that describe the PCI Root bridge
resources.

When ECAM base address in ACPI info is 0, _STA control method
reports the device as disabled.

Signed-off-by: Xenia Ragiadakou <xenia.ragiadakou@amd.com>
Reviewed-by: Jason Andryuk <jason.andryuk@amd.com>