libxl/pvh: add virtio-pci support
Introduce a new config option 'virtio_pci' to launch a PVH guest
with a QEMU emulated PCI Host bridge for virtio-pci devices.
The virtio-pci devices can be attached passing the appropriate
QEMU arguments to the config option 'device_model_args'.

The PCI Segment 1 resources, as defined in public/hvm/e820.h,
are passed to the QEMU instance using 'xenpvh' machine properties
to be used as virtio-pci bridge resources.

When virtio-pci support is enabled, the virtio-pci bridge ECAM
space is reserved in guest's E820, the ACPI info PCI1 fields
are populated to enable PCI bridge device in DSDT and an MCFG
table is added to the guest's static ACPI tables.

Signed-off-by: Xenia Ragiadakou <xenia.ragiadakou@amd.com>
Reviewed-by: Jason Andryuk <jason.andryuk@amd.com>