x86/pvh: enable ACPI Fixed Power and Sleep Button events
The support of ACPI fixed power and sleep button events is required
for triggering s2idle suspend and resume to a PVH guest and is
indicated in FADT by the respective flags being zero.

In order to support those events, ACPI PM emulation needs to be
enabled for PVH guests to emulate accesses to PM1 status, enable
and control registers.
The emulation of PM1 control registers is not symmetrical between
HVM and PVH guests because in the case of HVM guests those registers
are emulated by QEMU. For that reason, the presence of emulated VGA
was used to distinguish between HVM and PVH guests. Same way is used
in linux.
SCI_EN bit of PM1 control register indicates that the PM events
generate an SCI interrupt and needs to be always set.
Currently only S0 state is supported for PVH guests and _Sx objects
are not present in PVH ACPI DSDT. Therefore, accesses to SLP_TYPx
and SLP_EN bits of PM1 control registers are not expected to occur.
Hence, writes to PM1 control registers are ignored and reads
to other than SCI_EN bit return zero.

Since most guests use ACPI ioport location of version 1 and
for HVM guests the switch to version 1 is performed by the hvmloader,
for PVH guests version 1 location is used since the beginning.

GPE0_BLK needs to be disabled in the FADT of xl launched PVH guests
because it is not emulated, and when a PM event is triggered the
guest in order to assess the resource of the event will attempt to
access it.

Signed-off-by: Xenia Ragiadakou <xenia.ragiadakou@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>