xen/common: add grant table reset method
Shared and active entries are simply reset, avoiding the need for
reallocation. The maptracks tree is freed, as it would contain invalid
entries after reset.

To ensure consistency, all grant table entries are cleared. Keeping
references to granted pages from other domains could break the boot
process of a domain that expects no preconfigured grant tables. This
clearing will result in broken mapping references in the granted domain.

This is a design choice: it is currently assumed that other domains are
responsible for fixing or freeing any reset grants.

Signed-off-by: Luca Miccio <luca.miccio@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>