xen, dom0less-build: expose xenstore init function
Since a domain reset will clear all grant tables and event channels, it
is up to Xen to restore the initial status for xenstore. Dom0less
domains, which are the main target domains for the reset feature, can be
configured during boot with enhanced features such that they can use PV
drivers. The init function should be used by the reset method as well.
Expose the init function that will be later used in domain reset.

Signed-off-by: Luca Miccio <luca.miccio@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>