xen/arm: implement domain reset
This commit introduces support for resetting Dom0less domains on ARM
platforms. Dom0 and xl-based domains are currently not supported.

Boot modules for domains are stored in a dedicated memory region, whose
size can be configured via KConfig.

To avoid blocking the hypervisor for too long, the copy operation is
split into multiple preemptible chunks of 64 KiB. This size was chosen
as a practical trade-off: it is large enough to maintain good copy
throughput, while small enough to stay well below typical hypercall
preemption thresholds, minimizing latency and ensuring responsiveness.

During self-reset on arm64, the return value of the main hypercall must
be set to the GPA kernel entry. Otherwise, the domain will fail to boot
due to an invalid x0 register.

Signed-off-by: Luca Miccio <luca.miccio@amd.com>
Tested-by: Stewart Hildebrand <stewart.hildebrand@amd.com>
[stefano: code style; adding missing put_page]
Signed-off-by: Stefano Stabellini <stefano.stabellini@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>