tools/init-dom0less: add force option
When a domain is reset, its xenstore data is also reset. This means
that, for instance, the grant table references are no longer valid and
need to be recreated.
Add a simple option that will try to remove the domains passed as
arguments before actually creating them, thus forcing their creation.

Signed-off-by: Luca Miccio <luca.miccio@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>